feat(grpc): all grpc version API

+++ count-emission-service/internal/model/emission/emission_entity.go
@@ -4,7 +4,7 @@ import "time"
 
 type Emission struct {
 	Id            uint64
-	UserId        int8
+	UserId        int32
 	VehicleType   string
 	FuelType      string
 	DistanceKm    float64
@@ -14,25 +14,25 @@ type Emission struct {
 }
 
 type UserDailyEmission struct {
-	UserId             int8
+	UserId             int32
 	Date               string
 	TotalEmissionKgCo2 float64
 }
 
 type UserMonthlyEmissionDetail struct {
-	UserId             int8
+	UserId             int32
 	Month              string
 	TotalEmissionKgCo2 float64
 }
 
 type UserMonthlyEmission struct {
-	UserId                    int8
+	UserId                    int32
 	DailyEmissions            []UserDailyEmission
 	TotalEmissionMonthlyKgCo2 float64
 }
 
 type UserYearlyEmission struct {
-	UserId                   int8
+	UserId                   int32
 	MonthlyEmissions         []UserMonthlyEmissionDetail
 	TotalEmissionYearlyKgCo2 float64
 }
