+++ count-emission-service/internal/repository/internal_repo/emission_repository.go
@@ -28,7 +28,7 @@ func (emr *EmissionRepository) CreateUserEmission(ctx context.Context, req emiss
 	return nil
 }
 
-func (emr *EmissionRepository) GetUserDailyEmission(ctx context.Context, userId int8) (*emission.UserDailyEmission, error) {
+func (emr *EmissionRepository) GetUserDailyEmission(ctx context.Context, userId int32) (*emission.UserDailyEmission, error) {
 	var userEmission emission.UserDailyEmission
 	userEmission.UserId = userId
 	err := emr.DB.WithContext(ctx).
@@ -42,7 +42,7 @@ func (emr *EmissionRepository) GetUserDailyEmission(ctx context.Context, userId 
 	return &userEmission, nil
 }
 
-func (emr *EmissionRepository) GetUserMonthlyEmission(ctx context.Context, userId int8) (*emission.UserMonthlyEmission, error) {
+func (emr *EmissionRepository) GetUserMonthlyEmission(ctx context.Context, userId int32) (*emission.UserMonthlyEmission, error) {
 	var (
 		result              []emission.UserDailyEmission
 		monthlyTotal        float64
@@ -107,7 +107,7 @@ func (emr *EmissionRepository) GetUserMonthlyEmission(ctx context.Context, userI
 	return &userMonthlyEmission, nil
 }
 
-func (emr *EmissionRepository) GetUserYearlyEmission(ctx context.Context, userId int8) (*emission.UserYearlyEmission, error) {
+func (emr *EmissionRepository) GetUserYearlyEmission(ctx context.Context, userId int32) (*emission.UserYearlyEmission, error) {
 	var (
 		result             []emission.UserMonthlyEmissionDetail
 		yearlyTotal        float64
